+++ src/neptune/cli/status.py
@@ -62,7 +62,6 @@ class StatusRunner:
                 "\nWARNING: %s objects was skipped because they do not exist anymore.",
                 len(containers.not_found_containers),
             )
-
         if not any([containers.synced_containers, containers.unsynced_containers, containers.offline_containers]):
             logger.info("There are no Neptune objects in %s", path)
             sys.exit(1)