+++ src/neptune/internal/backends/neptune_backend_mock.py
@@ -23,6 +23,7 @@ from shutil import copyfile
 from typing import (
     Any,
     Dict,
+    Generator,
     Iterable,
     List,
     Optional,
@@ -131,6 +132,7 @@ from neptune.types.series.string_series import StringSeries
 from neptune.types.sets.string_set import StringSet
 from neptune.types.value import Value
 from neptune.types.value_visitor import ValueVisitor
+from neptune.typing import ProgressBarType
 
 Val = TypeVar("Val", bound=Value)
 
@@ -330,6 +332,7 @@ class NeptuneBackendMock(NeptuneBackend):
         container_type: ContainerType,
         path: List[str],
         destination: Optional[str] = None,
+        progress_bar: Optional[ProgressBarType] = None,
     ):
         run = self._get_container(container_id, container_type)
         value: File = run.get(path)
@@ -349,6 +352,7 @@ class NeptuneBackendMock(NeptuneBackend):
         container_type: ContainerType,
         path: List[str],
         destination: Optional[str] = None,
+        progress_bar: Optional[ProgressBarType] = None,
     ):
         run = self._get_container(container_id, container_type)
         source_file_set_value: FileSet = run.get(path)
@@ -486,6 +490,7 @@ class NeptuneBackendMock(NeptuneBackend):
         path: List[str],
         index: int,
         destination: str,
+        progress_bar: Optional[ProgressBarType],
     ):
         """Non relevant for backend"""
 
@@ -542,7 +547,11 @@ class NeptuneBackendMock(NeptuneBackend):
         types: Optional[Iterable[ContainerType]] = None,
         query: Optional[NQLQuery] = None,
         columns: Optional[Iterable[str]] = None,
-    ) -> List[LeaderboardEntry]:
+        limit: Optional[int] = None,
+        sort_by: str = "sys/creation_time",
+        ascending: bool = False,
+        progress_bar: Optional[ProgressBarType] = None,
+    ) -> Generator[LeaderboardEntry, None, None]:
         """Non relevant for mock"""
 
     class AttributeTypeConverterValueVisitor(ValueVisitor[AttributeType]):