+++ src/neptune/metadata_containers/metadata_containers_table.py
@@ -19,6 +19,7 @@ from datetime import datetime
 from typing import (
     Any,
     Dict,
+    Generator,
     List,
     Optional,
     Union,
@@ -38,6 +39,7 @@ from neptune.internal.utils.paths import (
     parse_path,
 )
 from neptune.internal.utils.run_state import RunState
+from neptune.typing import ProgressBarType
 
 logger = get_logger()
 
@@ -101,7 +103,12 @@ class TableEntry:
                 return None
         raise ValueError("Could not find {} attribute".format(path))
 
-    def download_file_attribute(self, path: str, destination: Optional[str]):
+    def download_file_attribute(
+        self,
+        path: str,
+        destination: Optional[str],
+        progress_bar: Optional[ProgressBarType] = None,
+    ):
         for attr in self._attributes:
             if attr.path == path:
                 _type = attr.type
@@ -111,12 +118,18 @@ class TableEntry:
                         container_type=self._container_type,
                         path=parse_path(path),
                         destination=destination,
+                        progress_bar=progress_bar,
                     )
                     return
                 raise MetadataInconsistency("Cannot download file from attribute of type {}".format(_type))
         raise ValueError("Could not find {} attribute".format(path))
 
-    def download_file_set_attribute(self, path: str, destination: Optional[str]):
+    def download_file_set_attribute(
+        self,
+        path: str,
+        destination: Optional[str],
+        progress_bar: Optional[ProgressBarType] = None,
+    ):
         for attr in self._attributes:
             if attr.path == path:
                 _type = attr.type
@@ -126,6 +139,7 @@ class TableEntry:
                         container_type=self._container_type,
                         path=parse_path(path),
                         destination=destination,
+                        progress_bar=progress_bar,
                     )
                     return
                 raise MetadataInconsistency("Cannot download ZIP archive from attribute of type {}".format(_type))
@@ -157,22 +171,28 @@ class Table:
         self,
         backend: NeptuneBackend,
         container_type: ContainerType,
-        entries: List[LeaderboardEntry],
+        entries: Generator[LeaderboardEntry, None, None],
     ):
         self._backend = backend
         self._entries = entries
         self._container_type = container_type
+        self._iterator = iter(entries if entries else ())
 
     def to_rows(self) -> List[TableEntry]:
-        return [
-            TableEntry(
-                backend=self._backend,
-                container_type=self._container_type,
-                _id=e.id,
-                attributes=e.attributes,
-            )
-            for e in self._entries
-        ]
+        return list(self)
+
+    def __iter__(self) -> "Table":
+        return self
+
+    def __next__(self) -> TableEntry:
+        entry = next(self._iterator)
+
+        return TableEntry(
+            backend=self._backend,
+            container_type=self._container_type,
+            _id=entry.id,
+            attributes=entry.attributes,
+        )
 
     def to_pandas(self):
         import pandas as pd